renamed energy contents
