+++ src/energy_contents.py
